+++ src/main/java/com/commercetools/sync/producttypes/utils/EnumUpdateActionUtils.java
@@ -26,7 +26,7 @@ import static java.util.Optional.of;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toMap;
 
-public final class EnumUpdateActionsUtils {
+public final class EnumUpdateActionUtils {
     /**
      * Given a list of new {@link EnumValue}s, gets a map where the keys are the enum value key, and the values
      * are the enum instances.
@@ -205,6 +205,6 @@ public final class EnumUpdateActionsUtils {
             .collect(Collectors.toList());
     }
 
-    private EnumUpdateActionsUtils() {
+    private EnumUpdateActionUtils() {
     }
 }
